backend/requirements.txt
flask~=2.0.1
flask-cors==3.0.10
pymongo~=3.12.0
okta_jwt_verifier~=0.1.0
flask[async]~=2.0.1
PyYAML~=5.4.1
Flask-SocketIO==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0
kafka-python==2.0.2
faker==8.10.3
werkzeug==2.0.1
pyjwt==2.1.0
